requirements.txt
# Agent Edge - Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
websockets>=12.0

# Google ADK + A2A
google-adk[a2a]
a2a-sdk
litellm

# MCP
mcp>=1.0.0
pyserial>=3.5

# Dashboard
streamlit>=1.29.0
streamlit-autorefresh>=1.0.1
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
